Implement SyncWorker
