+++ android/geowidget/build.gradle
@@ -114,6 +114,7 @@ dependencies {
 
     //Hilt - Dependency Injection
     implementation "com.google.dagger:hilt-android:$hiltVersion"
+    implementation "androidx.hilt:hilt-work:$hiltWorkVersion"
     kapt "com.google.dagger:hilt-compiler:$hiltVersion"
 
     implementation 'androidx.core:core-ktx:1.8.0'
